add try catch block
